+++ tests/testthat/test-PKNCA.options.R
@@ -66,6 +66,7 @@ test_that("PKNCA.options", {
       ),
       first.tmax = TRUE,
       allow.tmax.in.half.life = FALSE,
+      keep_interval_cols = NULL,
       min.hl.points = 3,
       min.span.ratio = 2,
       max.aucinf.pext = 20,